+++ latest/nginx/nginx.conf
@@ -1,5 +1,8 @@
 # Nginx configuration for Dhwani API Proxy
 
+# PID file location (writable by non-root user)
+pid /data/nginx.pid;
+
 # Worker processes and connections
 worker_processes auto;
 events {
@@ -20,7 +23,7 @@ http {
 
     # Server block for HTTP (Cloudflare handles HTTPS)
     server {
-        listen 80;
+        listen 8080;  # Non-privileged port
         server_name proxy.dwani.ai;  # Replace with your domain
 
         # Security headers
